requirements.txt
allure-pytest==2.12.0
httpx==0.23.3
jsonschema==4.17.3
pydantic==1.10.4
pytest==7.2.0
python-dotenv==0.21.0